refactor!: add shuffle utility methods and move Iterable methods in a dedicated source file
