+++ alchemist-implementationbase/src/main/kotlin/it/unibo/alchemist/model/util/RandomGeneratorExtension.kt
@@ -1,6 +1,6 @@
 /*
- * Copyright (C) 2010-2020, Danilo Pianini and contributors
- * listed in the main project's alchemist/build.gradle.kts file.
+ * Copyright (C) 2010-2022, Danilo Pianini and contributors
+ * listed, for each module, in the respective subproject's build.gradle.kts file.
  *
  * This file is part of Alchemist, and is distributed under the terms of the
  * GNU General Public License, with a linking exception,
@@ -9,7 +9,6 @@
 
 package it.unibo.alchemist.model.util
 
-import java.util.Collections
 import org.apache.commons.math3.random.RandomGenerator
 
 /**
@@ -25,18 +24,4 @@ object RandomGeneratorExtension {
      *          the upper bound.
      */
     fun RandomGenerator.nextDouble(from: Double, to: Double) = nextDouble() * (to - from) + from
-
-    /**
-     * [Fisher–Yates shuffle algorithm](https://bit.ly/33Z3xFu)
-     * using Apache random number generator.
-     * More information [on Wikipedia](https://en.wikipedia.org/wiki/Fisher-Yates_shuffle).
-     *
-     * @param randomGenerator
-     *          the simulation {@link RandomGenerator}.
-     */
-    fun <R> Iterable<R>.shuffled(randomGenerator: RandomGenerator): Iterable<R> = toMutableList().apply {
-        for (i in size - 1 downTo 1) {
-            Collections.swap(this, i, randomGenerator.nextInt(i + 1))
-        }
-    }
 }
